requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.3
pyalsaaudio==0.10.0
PyAudio==0.2.14
rapidfuzz==3.6.1
requests==2.31.0
scipy==1.12.0
soupsieve==2.5
syncedlyrics==0.8.0
urllib3==2.2.0
